requirements.txt
requests~=2.28.1
gspread~=5.4.0
oauth2client~=4.1.3
google-api-python-client~=2.57.0
discord.py~=1.7.3
python-dotenv~=0.20.0
attrs~=22.1.0
enum34~=1.1.10
numpy~=1.23.2
scipy~=1.9.1
construct~=2.10.68
wszst-yaz0~=1.2.0.post1
dropbox~=11.36.2